+++ pool/src/lib.rs
@@ -1,8 +1,18 @@
-use crate::guard::TaskType;
-use candid::{CandidType, Nat, Principal};
-use cycles_minting_canister::NotifyError;
-use ic_ledger_core::block::BlockType;
-use ic_types::Cycles;
+pub use crate::guard::{GuardPrincipal, TaskGuard, TaskType};
+pub use crate::memory::{
+    add_member_total_cycles, add_rewards, commit_block_participants, get_member_cycles,
+    get_miner_canister, get_next_block_participants, get_rewards, set_member_block_cycles,
+    set_miner_canister, set_rewards, total_pending_rewards,
+};
+pub use crate::system_calls::{fetch_block, notify_top_up, transfer};
+pub use crate::types::{MemberCycles, Rewards};
+
+mod guard;
+mod memory;
+mod system_calls;
+mod types;
+
+use candid::Principal;
 use std::cell::RefCell;
 use std::collections::BTreeSet;
 
@@ -20,73 +30,6 @@ pub const MAINNET_LEDGER_INDEX_CANISTER_ID: Principal =
 
 pub const MAINNET_CYCLE_MINTER_CANISTER_ID: Principal =
     Principal::from_slice(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x01, 0x01]);
-
-pub mod guard;
-pub mod memory;
-mod types;
-
-pub use crate::types::*;
-
-pub async fn fetch_block(block_height: u64) -> Result<icp_ledger::Block, String> {
-    let args = icrc_ledger_types::icrc3::blocks::GetBlocksRequest {
-        start: block_height.into(),
-        length: Nat::from(1_u8),
-    };
-
-    let res = ic_cdk::api::call::call::<_, (ic_icp_index::GetBlocksResponse,)>(
-        MAINNET_LEDGER_INDEX_CANISTER_ID,
-        "get_blocks",
-        (args,),
-    )
-    .await;
-    match res {
-        Ok(res) => {
-            if let Some(block_raw) = res.0.blocks.first() {
-                Ok(icp_ledger::Block::decode(block_raw.clone()).unwrap())
-            } else {
-                Err(format!(
-                    "Block {} not available in ICP index canister",
-                    block_height
-                ))
-            }
-        }
-        Err((code, msg)) => Err(format!(
-            "Error while calling ICP index canister ({:?}): {}",
-            code, msg
-        )),
-    }
-}
-
-#[derive(CandidType)]
-struct NotifyTopUp {
-    block_index: u64,
-    canister_id: Principal,
-}
-
-pub async fn notify_top_up(block_height: u64) -> Result<Cycles, String> {
-    let canister_id = ic_cdk::id();
-    let args = NotifyTopUp {
-        block_index: block_height,
-        canister_id,
-    };
-
-    let res = ic_cdk::api::call::call::<_, (Result<Cycles, NotifyError>,)>(
-        MAINNET_CYCLE_MINTER_CANISTER_ID,
-        "notify_top_up",
-        (args,),
-    )
-    .await;
-    match res {
-        Ok(res) => match res.0 {
-            Ok(cycles) => Ok(cycles),
-            Err(e) => Err(format!("Error from cycles minting canister: {e}")),
-        },
-        Err((code, msg)) => Err(format!(
-            "Error while calling cycles minting canister ({:?}): {}",
-            code, msg
-        )),
-    }
-}
 
 thread_local! {
     static __STATE: RefCell<State> = RefCell::default();
